Maintaining the variable resolver method along with the other methods.
