+++ src/Verso.Ado/Helpers/CredentialResolver.cs
@@ -10,6 +10,7 @@ namespace Verso.Ado.Helpers;
 internal static class CredentialResolver
 {
     private static readonly Regex EnvPattern = new(@"\$env:([A-Za-z_][A-Za-z0-9_]*)", RegexOptions.Compiled);
+    private static readonly Regex VarPattern = new(@"\$var:([A-Za-z_][A-Za-z0-9_]*)", RegexOptions.Compiled);
     private static readonly Regex SecretPattern = new(@"\$secret:([A-Za-z_][A-Za-z0-9_]*)", RegexOptions.Compiled);
     private static readonly Regex PasswordPattern = new(
         @"(Password|Pwd)\s*=\s*([^;]*)",
@@ -35,7 +36,7 @@ internal static class CredentialResolver
         }
 
         // Expand $var: tokens from the notebook variable store
-        var (varResolved, varError) = VariablePlaceholderResolver.Resolve(raw, variables, "connection string");
+        var (varResolved, varError) = ResolveVariable(raw, variables, "connection string");
         if (varError is not null)
             return (null, varError);
 
@@ -60,6 +61,41 @@ internal static class CredentialResolver
         return (resolved, null);
     }
 
+    internal static (string? Resolved, string? Error) ResolveVariable(string raw, IVariableStore? variables, string targetName)
+    {
+        if (!VarPattern.IsMatch(raw))
+            return (raw, null);
+
+        if (variables is null)
+            return (null, "$var: placeholders require a variable store but none is available.");
+
+        string? error = null;
+        var resolved = VarPattern.Replace(raw, match =>
+        {
+            var varName = match.Groups[1].Value;
+            var allVars = variables.GetAll();
+            var descriptor = allVars.FirstOrDefault(v =>
+                string.Equals(v.Name, varName, StringComparison.OrdinalIgnoreCase));
+
+            if (descriptor is null)
+            {
+                error = $"Variable '{varName}' is not defined. Set it in a C# cell before using $var:{varName}.";
+                return match.Value;
+            }
+
+            var value = descriptor.Value?.ToString();
+            if (string.IsNullOrEmpty(value))
+            {
+                error = $"Variable '{varName}' is null or empty for {targetName}.";
+                return match.Value;
+            }
+
+            return value;
+        });
+
+        return error is null ? (resolved, null) : (null, error);
+    }
+    
     /// <summary>
     /// Replaces <c>Password=...</c> and <c>Pwd=...</c> values with <c>***</c> for safe display.
     /// </summary>
